Pull invalid length hint tests into a separate file
